Move & clean aurelia
