--- app/aurelia/src/demo/index.ts
@@ -1,2 +1,0 @@
-export { default as Button } from './button';
-export { default as Welcome } from './welcome';
