+++ lib/cli/src/frameworks/aurelia/welcome.ts
