+++ lib/cli/src/generators/AURELIA/index.ts
@@ -1,15 +1,6 @@
-import {
-  getVersionedPackages,
-  writePackageJson,
-  getBabelDependencies,
-  installDependencies,
-  getPackageJson,
-  writeFileAsJson,
-  copyTemplate,
-  readFileAsJson,
-} from '../../helpers';
-import { Generator } from '../generator';
+import { writeFileAsJson, readFileAsJson } from '../../helpers';
 import { StoryFormat } from '../../project_types';
+import baseGenerator, { Generator } from '../generator';
 
 function addStorybookExcludeGlobToTsConfig() {
   const tsConfigJson = readFileAsJson('tsconfig.json', true);
@@ -26,36 +17,13 @@ function addStorybookExcludeGlobToTsConfig() {
   tsConfigJson.exclude = [...exclude, glob];
   writeFileAsJson('tsconfig.json', tsConfigJson);
 }
-const generator: Generator = async (npmOptions, { storyFormat = StoryFormat.CSF }) => {
-  copyTemplate(__dirname, storyFormat);
-  const packages = [
-    '@storybook/aurelia',
-    '@storybook/addon-actions',
-    '@storybook/addon-links',
-    '@storybook/addons',
-    '@storybook/addon-storysource',
-    '@storybook/addon-knobs',
-    '@storybook/addon-options',
-    '@storybook/addon-a11y',
-    '@storybook/addon-backgrounds',
-    'aurelia',
-  ];
 
-  if (storyFormat === 'mdx') {
-    packages.push('@storybook/addon-docs');
-  }
-
-  const versionedPackages = await getVersionedPackages(npmOptions, ...packages);
-  const packageJson = getPackageJson();
-  packageJson.dependencies = packageJson.dependencies || {};
-  packageJson.devDependencies = packageJson.devDependencies || {};
-  packageJson.scripts = packageJson.scripts || {};
-  packageJson.scripts.storybook = 'start-storybook -p 6006';
-  packageJson.scripts['build-storybook'] = 'build-storybook';
-  writePackageJson(packageJson);
+const generator: Generator = async (npmOptions, options) => {
   addStorybookExcludeGlobToTsConfig();
-  const babelDependencies = await getBabelDependencies(npmOptions, packageJson);
-  installDependencies({ ...npmOptions, packageJson }, [...versionedPackages, ...babelDependencies]);
+  baseGenerator(npmOptions, options, 'aurelia', {
+    dirname: options.storyFormat === StoryFormat.MDX ? __dirname : undefined,
+    extraPackages: ['aurelia'],
+  });
 };
 
 export default generator;
