--- lib/cli/src/generators/AURELIA/template-csf/.storybook/main.js
@@ -1,4 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.ts'],
-  addons: ['@storybook/addon-actions', '@storybook/addon-links'],
-};
