--- lib/cli/src/generators/AURELIA/template-mdx/.storybook/main.js
@@ -1,4 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.(ts|mdx)'],
-  addons: ['@storybook/addon-docs', '@storybook/addon-actions', '@storybook/addon-links'],
-};
