+++ lib/cli/src/generators/AURELIA/template-mdx/src/stories/0-Welcome.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
-import { Welcome } from '@storybook/aurelia/demo';
+import Welcome from './welcome';
 
 <Meta title="Welcome" component={Welcome} />
