+++ lib/cli/src/generators/AURELIA/template-mdx/src/stories/1-Button.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
 import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 
-import { Button } from '@storybook/aurelia/demo';
+import Button from './button';
 
 <Meta title="Button" component={Button} />
